+++ Archive-Extraction-tool/Archive-Extraction-tool/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace Unzipper
+namespace Archive_Extraction_tool
 {
     /// <summary>
     /// Interaction logic for App.xaml