+++ Changelog.md
@@ -1,5 +1,12 @@
+V2.0: 
+.net5.0
+Ookii.Dialogs.Wpf				V3.1.0
+Serilog							V2.10.0
+Serilog.Settings.AppSettings	V2.2.2
+SharpCompress					V0.28.3
+
 Initial model
 
 .net 4.8
-Ookii.Dialogs.Wpf 3.1.0
-System.IO.Compression.ZipFile 4.3.0
+Ookii.Dialogs.Wpf 				3.1.0
+System.IO.Compression.ZipFile 	4.3.0